Made types implement Firestore types and add package metadata
The biggest change in this commit is making the exported classes (e.g.,
`QueryWrapper`, `DocumentWrapper`, etc.) implement the base types from Firestore
to ensure they expose all the same methods as a normal Firestore object. This
makes it easier to use the wrappers instead of having to use the `.ref` property
to access the raw reference object. All the missing methods and properties have
been added onto the classes to implement those interfaces properly.

This commit also adds better type support for unions of types. The previous
classes assumed the generic types were always a single object (e.g., `Document
extends GenericFirestoreDocument` matches a single document, but it can also
match a union of multiple documents, all of which need to be handled properly).
In general, you don't have to worry about whether the generic type is a single
object or a union of multiple types since TypeScript handles them as a unit in
most cases, but there were some issues with `keyof`. For example `{ foo: string,
bar: number } | { foo: string, qux: number }` can be considered as extending `{
[key: string]: any }`, but if you perform `keyof` on the union type, instead of
getting `"foo" | "bar" | "qux"` (i.e., all the keys in all the objects),
TypeScript returns keys that are part of *all* objects, not *any*, so the return
type is just `"foo"` since that's the only key on both objects. Since the schema
types can be a union of multiple documents/collections, some of the types were
updated to support unions of objects (e.g., `DocumentsIn<Collection>` returns
all document objects in a collection, or union of collections, even if not all
the collections in the union have the same set of documents). Some other methods
also had their types improved by not assuming the generic type is a union of
types instead of a single type.

I also added more data to the `package.json` to prepare for publishing the
package.