Added missing FirestoreWrapper methods
The FirestoreWrapper class now implements all the methods of the underlying
Firestore object, which makes it possible to use the FirestoreWrapper as a
replacement for the normal Firestore object. All the same methods are
implemented with the same behavior so that the FirestoreWrapper can be used on
its own.

The `runTransaction` and `batch` methods of the `FirestoreWrapper` class also
needed to have their own wrapper classes set up (`ReadOnlyTransactionWrapper`,
`ReadWriteTransactionWrapper`, and `BatchWriteWrapper`) to allow those methods
to be used with `DocumentWrapper` objects (and the other variations like
`CollectionWrapper`, etc.). This makes it so that if the user were to use either
of those two methods, they wouldn't have to pass the raw `.ref`erence of each
wrapper class and can instead pass the wrapper itself. These three new wrapper
classes also are made to be interchangeable with their unwrapped counterparts.
This essentially makes the entire API for this library the same as the
underlying Firestore API, which should help with using the library so that the
user can use the same methods they're familiar with when using the normal
Firestore API.